Restrict dependencies and update Python version
This commit changes the `requires-python` field to exact version 3.11.9 and removes support for Python 3.12 and 3.13 related packages. It also refines dependency versions to be exact and removes the `loguru` package along with its dependency `win32-setctime`, simplifying the dependency list. These changes aim to ensure compatibility and simplify dependency management.